refactor(e2e): split team members tests into separate files
- Split monolithic 02-02-03-team-members.test.ts into 2 files:
  - 02-02-03-01-team-members-add-created.test.ts
  - 02-02-03-02-team-members-add-specific.test.ts
- Use shared team helpers for user selection and dialog management
